Machine learning Models along with Text Analytics
Part 1: Machine Learning Models

You work for an office transport company. You are in discussions with ABC Consulting company for 
providing transport for their employees. For this purpose, you are tasked with understanding how the 
employees of ABC Consulting prefer to commute presently (between home and office). Based on the 
parameters like age, salary, work experience, etc. given in the data set ‘Transport.csv’, you are required to 
predict the preferred mode of transport. The project requires you to build several Machine Learning 
models and compare them so that the model can be finalized.

Part 2: Text Mining 

A dataset of Shark Tank episodes is made available. It contains 495 entrepreneurs making their pitch 
to the VC sharks.